Optimize glyph rendering
This commit reduces the data throughput for every rendered glyph by
normalizing UVs and colors to fixed-point values and lowering
precision. Previously, a single color struct had a whopping size of 16
bytes, and UVs clocked in at 8 bytes. Now they both size at 4 bytes.

See #20

+++ src/graphics/renderer.h
@@ -16,7 +16,7 @@ struct graphics_font {
 };
 
 struct graphics_color {
-    float r, g, b, a;
+    uint8_t r, g, b, a;
 };
 
 struct graphics_position {
